Separate Projection class from ProjDiffAntiDrift
ProjDiffAntiDrift contains a GUI component. It is obvious to separate it from the logic part.
